+++ apps/monitoring_dash_app.py
@@ -90,4 +90,4 @@ def _trend(metric, _):
 
 
 if __name__ == "__main__":
-    app.run_server(debug=True, port=8055)
+    app.run(debug=True, port=8055)